refactor(email_api): splitter main.rs (1445→650 LOC) en 5 sous-modules
- event_bus.rs (197 LOC): MailEvent, persist/emit, api_events, api_events_stream
- deliverability_dto.rs (27 LOC): DTOs diagnostics/checklist/reminder
- mailing_list.rs (129 LOC): MailingListRequest + create/send
- dkim_service.rs (52 LOC): impl DkimService for RealDkimService
- main_tests.rs (412 LOC): mod tests déporté

Aucun changement de comportement.
