+++ src/bin/email_api_dir/mailing_list.rs
@@ -1,0 +1,129 @@
+use super::*;
+
+#[derive(Deserialize)]
+pub(super) struct MailingListRequest {
+    label: String,
+    emails: Vec<String>,
+}
+
+#[derive(Deserialize)]
+pub(super) struct MailingListEmailRequest {
+    from: String,
+    subject: String,
+    body: String,
+    mailing_list: String,
+}
+
+pub(super) async fn create_mailing_list(mailing_list: web::Json<MailingListRequest>) -> impl Responder {
+    let mailing_list_dir = Path::new("mailing-lists");
+    if !mailing_list_dir.exists() {
+        if let Err(e) = create_dir_all(mailing_list_dir) {
+            return HttpResponse::InternalServerError().json(serde_json::json!({
+                "status": "error",
+                "message": format!("Failed to create mailing list directory: {}", e)
+            }));
+        }
+    }
+
+    let file_path = mailing_list_dir.join(format!("{}.csv", mailing_list.label));
+    let mut file = match File::create(&file_path) {
+        Ok(file) => file,
+        Err(e) => {
+            return HttpResponse::InternalServerError().json(serde_json::json!({
+                "status": "error",
+                "message": format!("Failed to create file: {}", e)
+            }))
+        }
+    };
+
+    for email in &mailing_list.emails {
+        if let Err(e) = writeln!(file, "{}", email) {
+            return HttpResponse::InternalServerError().json(serde_json::json!({
+                "status": "error",
+                "message": format!("Failed to write to file: {}", e)
+            }));
+        }
+    }
+
+    HttpResponse::Ok().json(serde_json::json!({
+        "status": "success",
+        "message": format!("Mailing list '{}' created successfully", mailing_list.label)
+    }))
+}
+
+pub(super) async fn send_to_mailing_list(email_req: web::Json<MailingListEmailRequest>) -> impl Responder {
+    let mailing_list_path =
+        Path::new("mailing-lists").join(format!("{}.csv", email_req.mailing_list));
+
+    if !mailing_list_path.exists() {
+        return HttpResponse::NotFound().json(serde_json::json!({
+            "status": "error",
+            "message": format!("Mailing list '{}' not found", email_req.mailing_list)
+        }));
+    }
+
+    let file = match File::open(&mailing_list_path) {
+        Ok(file) => file,
+        Err(e) => {
+            return HttpResponse::InternalServerError().json(serde_json::json!({
+                "status": "error",
+                "message": format!("Failed to open mailing list file: {}", e)
+            }))
+        }
+    };
+
+    let reader = BufReader::new(file);
+    let mut success_count = 0;
+    let mut failure_count = 0;
+
+    for line in reader.lines() {
+        let to_email = match line {
+            Ok(email) => email.trim().to_string(),
+            Err(_e) => {
+                failure_count += 1;
+                continue;
+            }
+        };
+
+        let from = email_req.from.clone();
+        let to = to_email;
+        let subject = email_req.subject.clone();
+        let body = email_req.body.clone();
+
+        let _email_content = format!(
+            "From: {}\r\nTo: {}\r\nSubject: {}\r\n\r\n{}",
+            from, to, subject, body
+        );
+        let client = reqwest::Client::new();
+        let dkim_service_url = env::var("DKIM_SERVICE_URL").expect("DKIM_SERVICE_URL not set");
+
+        let _dkim_response = match client
+            .post(&dkim_service_url)
+            .json(&serde_json::json!({
+                "from": from,
+                "to": to,
+                "subject": subject,
+                "text": body
+            }))
+            .send()
+            .await
+        {
+            Ok(resp) => {
+                if resp.status().is_success() {
+                    success_count += 1;
+                } else {
+                    failure_count += 1;
+                }
+            }
+            Err(_) => {
+                failure_count += 1;
+            }
+        };
+    }
+
+    HttpResponse::Ok().json(serde_json::json!({
+        "status": "success",
+        "message": format!("Emails sent to mailing list '{}'. Successful: {}, Failed: {}",
+                           email_req.mailing_list, success_count, failure_count)
+    }))
+}
