+++ src/bin/email_api_dir/main.rs
@@ -55,6 +55,15 @@ mod admin_ops;
 mod external_handlers;
 mod external_probe_handlers;
 mod helpers;
+mod event_bus;
+mod deliverability_dto;
+mod mailing_list;
+mod dkim_service;
+#[cfg(test)] mod main_tests;
+pub use event_bus::*;
+pub use deliverability_dto::*;
+pub use mailing_list::*;
+pub use dkim_service::*;
 use helpers::{normalize_segment, build_misfits_local, normalize_oauth_provider, req_ip_str, get_accept_language, welcome_email_html};
 
 pub use auth_handlers::*;
@@ -93,201 +102,7 @@ use tokio::sync::broadcast;
 use uuid::Uuid;
 
 // --- Mail event monitoring ---
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-enum MailEventKind {
-    Sent,
-    Received,
-    Read,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct MailEvent {
-    id: String,
-    kind: MailEventKind,
-    user_id: String,
-    email_id: String,
-    subject: String,
-    from: String,
-    to: String,
-    timestamp: String,
-}
-
-type EventBus = broadcast::Sender<MailEvent>;
-
-// --- Shared auth types (moved to auth_handlers.rs) ---
-// LoginRequest, RegisterRequest, OAuthCallbackQuery, TwoFactorVerifyRequest
-// PasswordResetRequestBody, PasswordResetConfirmBody are in auth_handlers module.
-
-// --- 2FA default helper (still needed by main.rs serde defaults) ---
-
-// --- Send queue types ---
-
-const SEND_QUEUE_COLL: &str = "send_queue";
-
-#[derive(Deserialize)]
-struct UndoSendRequest {
-    id: String,
-}
-
-#[derive(Deserialize)]
-struct ScheduleSendBody {
-    #[serde(default)]
-    to: Vec<ComposerRecipient>,
-    #[serde(default)]
-    cc: Vec<ComposerRecipient>,
-    #[serde(default)]
-    bcc: Vec<ComposerRecipient>,
-    #[serde(default)]
-    subject: String,
-    #[serde(default)]
-    body: String,
-    #[serde(default)]
-    attachments: Vec<ComposeAttachmentInput>,
-    #[serde(default)]
-    from: Option<String>,
-    #[serde(default, rename = "inReplyTo", alias = "in_reply_to")]
-    in_reply_to: Option<String>,
-    #[serde(default)]
-    references: Vec<String>,
-    send_at: String,
-}
-
-// --- OAuth provider response types ---
-
-#[derive(Deserialize)]
-struct GithubTokenResponse {
-    access_token: Option<String>,
-    error: Option<String>,
-    error_description: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct GithubUser {
-    id: u64,
-    login: String,
-    name: Option<String>,
-    email: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct GithubEmail {
-    email: String,
-    primary: bool,
-    verified: bool,
-}
-
-#[derive(Serialize)]
-struct UserResponse {
-    id: String,
-    email: String,
-    display_name: String,
-    role: String,
-    two_factor_enabled: bool,
-    created_at: String,
-    updated_at: String,
-}
-
-#[derive(Serialize)]
-struct SessionResponse {
-    id: String,
-    user: UserResponse,
-    access_token: String,
-    refresh_token: String,
-    expires_at: u64,
-    refresh_expires_at: u64,
-    issued_at: u64,
-}
-
-#[derive(Serialize)]
-struct AuthResponse {
-    session: SessionResponse,
-}
-
-// make_session, verify_totp, generate_totp_secret, generate_otp_code
-// → moved to auth_handlers module
-
-
-
-async fn persist_event(mongo: &mongodb::Client, event: &MailEvent) {
-    let db = env::var("MONGODB_DATABASE").unwrap_or_else(|_| "mailserver".to_string());
-    let coll = mongo
-        .database(&db)
-        .collection::<bson::Document>("mail_events");
-    if let Ok(doc) = bson::to_document(event) {
-        if let Err(e) = coll.insert_one(doc).await {
-            eprintln!("persist_event error: {}", e);
-        }
-    }
-}
-
-async fn emit_event(bus: &EventBus, mongo: &mongodb::Client, event: MailEvent) {
-    persist_event(mongo, &event).await;
-    let _ = bus.send(event);
-}
-
-async fn api_events(
-    req: actix_web::HttpRequest,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let user_id = resolve_user_id(&req);
-    let db = env::var("MONGODB_DATABASE").unwrap_or_else(|_| "mailserver".to_string());
-    let coll = mongo
-        .database(&db)
-        .collection::<bson::Document>("mail_events");
-    let filter = doc! { "user_id": &user_id };
-    match coll
-        .find(filter)
-        .sort(doc! { "timestamp": -1 })
-        .limit(200)
-        .await
-    {
-        Ok(cursor) => match cursor.try_collect::<Vec<_>>().await {
-            Ok(docs) => {
-                let events: Vec<serde_json::Value> = docs
-                    .into_iter()
-                    .filter_map(|d| bson::from_document::<MailEvent>(d).ok())
-                    .filter_map(|e| serde_json::to_value(&e).ok())
-                    .collect();
-                HttpResponse::Ok().json(serde_json::json!({ "events": events }))
-            }
-            Err(e) => HttpResponse::InternalServerError()
-                .json(serde_json::json!({ "error": e.to_string() })),
-        },
-        Err(e) => {
-            HttpResponse::InternalServerError().json(serde_json::json!({ "error": e.to_string() }))
-        }
-    }
-}
-
-async fn api_events_stream(bus: web::Data<EventBus>, req: actix_web::HttpRequest) -> HttpResponse {
-    let user_id = resolve_user_id(&req);
-    let rx = bus.subscribe();
-    let event_stream = stream::unfold((rx, user_id), |(mut rx, uid)| async move {
-        loop {
-            match rx.recv().await {
-                Ok(event) if event.user_id == uid => {
-                    let data = serde_json::to_string(&event).unwrap_or_default();
-                    let chunk = format!("data: {}\n\n", data);
-                    return Some((
-                        Ok::<web::Bytes, actix_web::Error>(web::Bytes::from(chunk)),
-                        (rx, uid),
-                    ));
-                }
-                Ok(_) => continue,
-                Err(broadcast::error::RecvError::Lagged(_)) => continue,
-                Err(broadcast::error::RecvError::Closed) => return None,
-            }
-        }
-    });
-    HttpResponse::Ok()
-        .content_type("text/event-stream")
-        .insert_header(("Cache-Control", "no-cache"))
-        .insert_header(("X-Accel-Buffering", "no"))
-        .streaming(event_stream)
-}
+// event bus + auth DTOs → event_bus.rs
 
 // ===========================================================================
 // SMTP Monitoring endpoints
@@ -308,31 +123,7 @@ use simple_smtp_server::security;
 
 
 
-#[derive(Deserialize)]
-struct DeliverabilityDiagnosticsQuery {
-    #[serde(default = "default_window")]
-    window: String,
-    domain: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct DeliverabilityProcedureUpdateRequest {
-    checklist: Option<Vec<DeliverabilityChecklistUpdate>>,
-    reminder: Option<DeliverabilityReminderUpdate>,
-}
-
-#[derive(Deserialize)]
-struct DeliverabilityChecklistUpdate {
-    id: String,
-    checked: bool,
-    note: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct DeliverabilityReminderUpdate {
-    enabled: bool,
-    cadence_hours: u32,
-}
+// deliverability DTOs → deliverability_dto.rs
 
 /// GET /api/monitoring/summary?window=15m
 // monitoring + security handlers → monitoring_handlers module
@@ -346,133 +137,7 @@ pub struct EmailRequest {
     body: String,
 }
 
-#[derive(Deserialize)]
-struct MailingListRequest {
-    label: String,
-    emails: Vec<String>,
-}
-
-#[derive(Deserialize)]
-struct MailingListEmailRequest {
-    from: String,
-    subject: String,
-    body: String,
-    mailing_list: String,
-}
-
-async fn create_mailing_list(mailing_list: web::Json<MailingListRequest>) -> impl Responder {
-    let mailing_list_dir = Path::new("mailing-lists");
-    if !mailing_list_dir.exists() {
-        if let Err(e) = create_dir_all(mailing_list_dir) {
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": format!("Failed to create mailing list directory: {}", e)
-            }));
-        }
-    }
-
-    let file_path = mailing_list_dir.join(format!("{}.csv", mailing_list.label));
-    let mut file = match File::create(&file_path) {
-        Ok(file) => file,
-        Err(e) => {
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": format!("Failed to create file: {}", e)
-            }))
-        }
-    };
-
-    for email in &mailing_list.emails {
-        if let Err(e) = writeln!(file, "{}", email) {
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": format!("Failed to write to file: {}", e)
-            }));
-        }
-    }
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "status": "success",
-        "message": format!("Mailing list '{}' created successfully", mailing_list.label)
-    }))
-}
-
-async fn send_to_mailing_list(email_req: web::Json<MailingListEmailRequest>) -> impl Responder {
-    let mailing_list_path =
-        Path::new("mailing-lists").join(format!("{}.csv", email_req.mailing_list));
-
-    if !mailing_list_path.exists() {
-        return HttpResponse::NotFound().json(serde_json::json!({
-            "status": "error",
-            "message": format!("Mailing list '{}' not found", email_req.mailing_list)
-        }));
-    }
-
-    let file = match File::open(&mailing_list_path) {
-        Ok(file) => file,
-        Err(e) => {
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": format!("Failed to open mailing list file: {}", e)
-            }))
-        }
-    };
-
-    let reader = BufReader::new(file);
-    let mut success_count = 0;
-    let mut failure_count = 0;
-
-    for line in reader.lines() {
-        let to_email = match line {
-            Ok(email) => email.trim().to_string(),
-            Err(_e) => {
-                failure_count += 1;
-                continue;
-            }
-        };
-
-        let from = email_req.from.clone();
-        let to = to_email;
-        let subject = email_req.subject.clone();
-        let body = email_req.body.clone();
-
-        let _email_content = format!(
-            "From: {}\r\nTo: {}\r\nSubject: {}\r\n\r\n{}",
-            from, to, subject, body
-        );
-        let client = reqwest::Client::new();
-        let dkim_service_url = env::var("DKIM_SERVICE_URL").expect("DKIM_SERVICE_URL not set");
-
-        let _dkim_response = match client
-            .post(&dkim_service_url)
-            .json(&serde_json::json!({
-                "from": from,
-                "to": to,
-                "subject": subject,
-                "text": body
-            }))
-            .send()
-            .await
-        {
-            Ok(resp) => {
-                if resp.status().is_success() {
-                    success_count += 1;
-                } else {
-                    failure_count += 1;
-                }
-            }
-            Err(_) => {
-                failure_count += 1;
-            }
-        };
-    }
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "status": "success",
-        "message": format!("Emails sent to mailing list '{}'. Successful: {}, Failed: {}",
-                           email_req.mailing_list, success_count, failure_count)
-    }))
-}
+// mailing list → mailing_list.rs
 
 async fn send_email_handler(
     email_req: web::Json<EmailRequest>,
@@ -980,466 +645,6 @@ async fn main() -> std::io::Result<()> {
     .await
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use actix_web::{test, App};
-    use dotenv::dotenv;
-    use mockall::mock;
-    use mockall::predicate::eq;
+// tests → main_tests.rs
 
-    mock! {
-        pub DkimService {
-            pub async fn sign_email(&self, email: &EmailRequest) -> Result<serde_json::Value, std::io::Error>;
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl DkimService for MockDkimService {
-        async fn sign_email(
-            &self,
-            email: &EmailRequest,
-        ) -> Result<serde_json::Value, std::io::Error> {
-            self.sign_email(email).await
-        }
-    }
-
-    #[actix_web::test]
-    async fn test_send_email() {
-        dotenv::from_filename(".env.test").ok();
-
-        let mut mock_dkim_service = MockDkimService::new();
-        mock_dkim_service
-            .expect_sign_email()
-            .with(eq(EmailRequest {
-                from: "sender@example.com".to_string(),
-                to: "recipient@example.com".to_string(),
-                subject: "Test Email".to_string(),
-                body: "This is a test email.".to_string(),
-            }))
-            .times(1)
-            .returning(|_| Ok(serde_json::json!({"status": "success", "messageId": "12345"})));
-
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(
-                    Box::new(mock_dkim_service) as Box<dyn DkimService>
-                ))
-                .route("/send-email", web::post().to(send_email_handler)),
-        )
-        .await;
-
-        let email_request = EmailRequest {
-            from: "sender@example.com".to_string(),
-            to: "recipient@example.com".to_string(),
-            subject: "Test Email".to_string(),
-            body: "This is a test email.".to_string(),
-        };
-        println!("Sending test request to /send-email");
-
-        let req = test::TestRequest::post()
-            .uri("/send-email")
-            .set_json(&email_request)
-            .to_request();
-        let resp = test::call_service(&app, req).await;
-        println!("Response status: {:?}", resp.status());
-        assert!(resp.status().is_success());
-    }
-
-    #[actix_web::test]
-    async fn test_api_hermes_chat_requires_messages() {
-        let app = test::init_service(
-            App::new().route("/api/hermes/chat", web::post().to(api_hermes_chat)),
-        )
-        .await;
-
-        let req = test::TestRequest::post()
-            .uri("/api/hermes/chat")
-            .set_json(serde_json::json!({ "messages": [] }))
-            .to_request();
-
-        let resp = test::call_service(&app, req).await;
-        assert_eq!(resp.status(), actix_web::http::StatusCode::BAD_REQUEST);
-    }
-
-    #[test]
-    fn test_hermes_chat_request_accepts_explicit_session_overrides() {
-        let parsed: HermesChatProxyRequest = serde_json::from_value(serde_json::json!({
-            "messages": [{"role":"user","content":"hello"}],
-            "threadId": "thread-123",
-            "userId": "admin",
-            "sessionId": "mail-thread-explicit",
-            "sessionKey": "user-explicit",
-            "maxTokens": 1200
-        }))
-        .expect("HermesChatProxyRequest should deserialize");
-
-        assert_eq!(parsed.thread_id.as_deref(), Some("thread-123"));
-        assert_eq!(parsed.user_id.as_deref(), Some("admin"));
-        assert_eq!(parsed.session_id.as_deref(), Some("mail-thread-explicit"));
-        assert_eq!(parsed.session_key.as_deref(), Some("user-explicit"));
-        assert_eq!(parsed.max_tokens, Some(1200));
-    }
-    #[actix_web::test]
-    async fn test_api_hermes_runs_requires_input() {
-        let app = test::init_service(
-            App::new().route("/api/hermes/runs", web::post().to(api_hermes_runs)),
-        )
-        .await;
-
-        let req = test::TestRequest::post()
-            .uri("/api/hermes/runs")
-            .set_json(serde_json::json!({}))
-            .to_request();
-
-        let resp = test::call_service(&app, req).await;
-        assert_eq!(resp.status(), actix_web::http::StatusCode::BAD_REQUEST);
-    }
-
-    #[test]
-    fn test_hermes_runs_request_accepts_explicit_session_overrides() {
-        let parsed: HermesRunsProxyRequest = serde_json::from_value(serde_json::json!({
-            "input": [{"role":"user","content":"hello"}],
-            "threadId": "thread-123",
-            "userId": "admin",
-            "sessionId": "mail-thread-explicit",
-            "sessionKey": "user-explicit",
-            "model": "hermes-agent"
-        }))
-        .expect("HermesRunsProxyRequest should deserialize");
-
-        assert!(parsed.input.is_some());
-        assert_eq!(parsed.thread_id.as_deref(), Some("thread-123"));
-        assert_eq!(parsed.user_id.as_deref(), Some("admin"));
-        assert_eq!(parsed.session_id.as_deref(), Some("mail-thread-explicit"));
-        assert_eq!(parsed.session_key.as_deref(), Some("user-explicit"));
-        assert_eq!(parsed.model.as_deref(), Some("hermes-agent"));
-    }
-
-    #[test]
-    fn test_normalize_hermes_base_url_strips_v1_and_slashes() {
-        assert_eq!(
-            normalize_hermes_base_url("http://172.16.12.2:8642/v1/"),
-            "http://172.16.12.2:8642"
-        );
-        assert_eq!(
-            normalize_hermes_base_url("http://172.16.12.2:8642/v1"),
-            "http://172.16.12.2:8642"
-        );
-        assert_eq!(
-            normalize_hermes_base_url("http://172.16.12.2:8642"),
-            "http://172.16.12.2:8642"
-        );
-    }
-
-    // ─── Sprint 1: Integration tests on critical flows ────────────────────────
-    //
-    // These tests exercise the HTTP handler shape (status codes, response
-    // structure) without a live MongoDB or SMTP server. They verify:
-    //  - Correct 4xx on bad / missing input
-    //  - Response JSON shape on valid input (where the handler does not
-    //    require a DB connection at the validation stage)
-    //
-    // Flow 1: POST /api/auth/login — rejects empty password
-    #[actix_web::test]
-    async fn test_auth_login_rejects_empty_password() {
-        let app = test::init_service(
-            App::new()
-                .app_data(
-                    web::JsonConfig::default()
-                        .error_handler(|err, req| {
-                            let resp = actix_web::HttpResponse::BadRequest()
-                                .json(serde_json::json!({ "code": "INVALID_JSON", "message": err.to_string() }));
-                            actix_web::error::InternalError::from_response(err, resp).into()
-                        })
-                )
-                .route("/api/auth/login", web::post().to(auth_login)),
-        )
-        .await;
-
-        // Missing password field → JSON parse error → 400
-        let req = test::TestRequest::post()
-            .uri("/api/auth/login")
-            .set_json(serde_json::json!({ "email": "test@misfits.ai" }))
-            .to_request();
-        let resp = test::call_service(&app, req).await;
-        assert_eq!(resp.status(), actix_web::http::StatusCode::BAD_REQUEST);
-    }
-
-    // Flow 1b: POST /api/auth/login — correct shape → 200 or 401 (never 500)
-    #[actix_web::test]
-    async fn test_auth_login_valid_shape_returns_200_or_401() {
-        dotenv::from_filename(".env.test").ok();
-        use std::sync::Arc;
-        let mongo_url = std::env::var("MONGODB_URI").unwrap_or_else(|_| "mongodb://localhost:27017".to_string());
-        let mongo = match mongodb::Client::with_uri_str(&mongo_url).await {
-            Ok(c) => Arc::new(c),
-            Err(_) => return, // Skip if no test DB
-        };
-        let logic = web::Data::new(Arc::new(Logic::new(mongo.clone())));
-        let mongo_data = web::Data::new(mongo);
-        let bus_data = web::Data::new(EventBus::new());
-
-        let app = test::init_service(
-            App::new()
-                .app_data(logic)
-                .app_data(mongo_data)
-                .app_data(bus_data)
-                .route("/api/auth/login", web::post().to(auth_login)),
-        )
-        .await;
-
-        let req = test::TestRequest::post()
-            .uri("/api/auth/login")
-            .set_json(serde_json::json!({
-                "email": "nonexistent@misfits.ai",
-                "password": "wrongpassword"
-            }))
-            .to_request();
-        let resp = test::call_service(&app, req).await;
-        let status = resp.status().as_u16();
-        assert!(
-            status == 200 || status == 401,
-            "Expected 200 or 401, got {}",
-            status
-        );
-    }
-
-    // Flow 2: POST /api/auth/2fa/verify — rejects missing fields
-    #[actix_web::test]
-    async fn test_2fa_verify_rejects_missing_fields() {
-        let app = test::init_service(
-            App::new()
-                .app_data(
-                    web::JsonConfig::default().error_handler(|err, _req| {
-                        let resp = actix_web::HttpResponse::BadRequest()
-                            .json(serde_json::json!({ "verified": false, "error": err.to_string() }));
-                        actix_web::error::InternalError::from_response(err, resp).into()
-                    })
-                )
-                .route("/api/auth/2fa/verify", web::post().to(api_2fa_verify)),
-        )
-        .await;
-
-        // Missing code → 400
-        let req = test::TestRequest::post()
-            .uri("/api/auth/2fa/verify")
-            .set_json(serde_json::json!({ "email": "test@misfits.ai", "method": "email" }))
-            .to_request();
-        let resp = test::call_service(&app, req).await;
-        assert_eq!(resp.status(), actix_web::http::StatusCode::BAD_REQUEST);
-    }
-
-    // Flow 3: GET /api/emails — rejects request with no user identity (empty folder)
-    #[actix_web::test]
-    async fn test_api_emails_empty_folder_returns_empty_list() {
-        dotenv::from_filename(".env.test").ok();
-        use std::sync::Arc;
-        let mongo_url = std::env::var("MONGODB_URI").unwrap_or_else(|_| "mongodb://localhost:27017".to_string());
-        let mongo = match mongodb::Client::with_uri_str(&mongo_url).await {
-            Ok(c) => Arc::new(c),
-            Err(_) => return,
-        };
-        let logic = web::Data::new(Arc::new(Logic::new(mongo.clone())));
-
-        let app = test::init_service(
-            App::new()
-                .app_data(logic)
-                .route("/api/emails", web::get().to(api_emails)),
-        )
-        .await;
-
-        // No x-user-id header → resolve_user_id returns empty string → returns empty list
-        let req = test::TestRequest::get()
-            .uri("/api/emails?folder=inbox&page=1&page_size=5")
-            .to_request();
-        let resp = test::call_service(&app, req).await;
-        // Handler always returns 200 (empty list for unknown user)
-        assert_eq!(resp.status(), actix_web::http::StatusCode::OK);
-        let body: serde_json::Value = test::read_body_json(resp).await;
-        assert!(body["emails"].is_array(), "Expected emails array");
-    }
-
-    // Flow 4: POST /api/send — rejects missing recipients
-    #[actix_web::test]
-    async fn test_api_send_rejects_empty_recipients() {
-        dotenv::from_filename(".env.test").ok();
-        use std::sync::Arc;
-        let mongo_url = std::env::var("MONGODB_URI").unwrap_or_else(|_| "mongodb://localhost:27017".to_string());
-        let mongo = match mongodb::Client::with_uri_str(&mongo_url).await {
-            Ok(c) => Arc::new(c),
-            Err(_) => return,
-        };
-        let logic = web::Data::new(Arc::new(Logic::new(mongo.clone())));
-        let mongo_data = web::Data::new(mongo);
-        let bus_data = web::Data::new(EventBus::new());
-
-        let app = test::init_service(
-            App::new()
-                .app_data(logic)
-                .app_data(mongo_data)
-                .app_data(bus_data)
-                .route("/api/send", web::post().to(api_send)),
-        )
-        .await;
-
-        let req = test::TestRequest::post()
-            .uri("/api/send")
-            .set_json(serde_json::json!({
-                "to": [],
-                "subject": "Test",
-                "body": "Hello"
-            }))
-            .to_request();
-        let resp = test::call_service(&app, req).await;
-        assert_eq!(resp.status(), actix_web::http::StatusCode::BAD_REQUEST);
-        let body: serde_json::Value = test::read_body_json(resp).await;
-        assert_eq!(body["sent"], false);
-        assert!(body["message"].is_string());
-    }
-
-    // Flow 5: GET /api/monitoring/summary — returns required fields
-    #[actix_web::test]
-    async fn test_monitoring_summary_shape() {
-        dotenv::from_filename(".env.test").ok();
-        use std::sync::Arc;
-        let mongo_url = std::env::var("MONGODB_URI").unwrap_or_else(|_| "mongodb://localhost:27017".to_string());
-        let mongo = match mongodb::Client::with_uri_str(&mongo_url).await {
-            Ok(c) => Arc::new(c),
-            Err(_) => return,
-        };
-        let mongo_data = web::Data::new(mongo);
-
-        let app = test::init_service(
-            App::new()
-                .app_data(mongo_data)
-                .route("/api/monitoring/summary", web::get().to(api_monitoring_summary)),
-        )
-        .await;
-
-        let req = test::TestRequest::get()
-            .uri("/api/monitoring/summary?window=1h")
-            .to_request();
-        let resp = test::call_service(&app, req).await;
-        assert_eq!(resp.status(), actix_web::http::StatusCode::OK);
-        let body: serde_json::Value = test::read_body_json(resp).await;
-        // Shape assertions: all required fields present
-        assert!(body["total"].is_number(), "Expected total field");
-        assert!(body["deliveryRate"].is_number(), "Expected deliveryRate field");
-        assert!(body["bounceRate"].is_number(), "Expected bounceRate field");
-        assert!(body["byStatus"].is_object(), "Expected byStatus object");
-    }
-
-    // Flow 6: GET /api/admin/users — returns 200 with users array (RBAC flag off)
-    #[actix_web::test]
-    async fn test_admin_users_list_shape() {
-        dotenv::from_filename(".env.test").ok();
-        use std::sync::Arc;
-        let mongo_url = std::env::var("MONGODB_URI").unwrap_or_else(|_| "mongodb://localhost:27017".to_string());
-        let mongo = match mongodb::Client::with_uri_str(&mongo_url).await {
-            Ok(c) => Arc::new(c),
-            Err(_) => return,
-        };
-        let mongo_data = web::Data::new(mongo);
-
-        let app = test::init_service(
-            App::new()
-                .app_data(mongo_data)
-                .route("/api/admin/users", web::get().to(api_admin_users_list)),
-        )
-        .await;
-
-        let req = test::TestRequest::get()
-            .uri("/api/admin/users")
-            .to_request();
-        let resp = test::call_service(&app, req).await;
-        assert_eq!(resp.status(), actix_web::http::StatusCode::OK);
-        let body: serde_json::Value = test::read_body_json(resp).await;
-        assert!(body["users"].is_array(), "Expected users array");
-        assert!(body["generatedAt"].is_string(), "Expected generatedAt");
-    }
-
-    // Flow 6b: GET /api/admin/whoami — returns enforced flag
-    #[actix_web::test]
-    async fn test_admin_whoami_returns_enforced_flag() {
-        dotenv::from_filename(".env.test").ok();
-        use std::sync::Arc;
-        let mongo_url = std::env::var("MONGODB_URI").unwrap_or_else(|_| "mongodb://localhost:27017".to_string());
-        let mongo = match mongodb::Client::with_uri_str(&mongo_url).await {
-            Ok(c) => Arc::new(c),
-            Err(_) => return,
-        };
-        let mongo_data = web::Data::new(mongo);
-
-        let app = test::init_service(
-            App::new()
-                .app_data(mongo_data)
-                .route("/api/admin/whoami", web::get().to(api_admin_whoami)),
-        )
-        .await;
-
-        let req = test::TestRequest::get()
-            .uri("/api/admin/whoami")
-            .insert_header(("Authorization", "Bearer test-token"))
-            .to_request();
-        let resp = test::call_service(&app, req).await;
-        // Either 200 (flag off, system user) or 401 (flag on, token unknown)
-        let status = resp.status().as_u16();
-        assert!(
-            status == 200 || status == 401,
-            "Expected 200 or 401, got {}",
-            status
-        );
-    }
-} // end mod tests
-
-#[async_trait::async_trait]
-pub trait DkimService: Send + Sync {
-    async fn sign_email(&self, email: &EmailRequest) -> Result<serde_json::Value, std::io::Error>;
-}
-
-pub struct RealDkimService;
-
-#[async_trait::async_trait]
-impl DkimService for RealDkimService {
-    async fn sign_email(&self, email: &EmailRequest) -> Result<serde_json::Value, std::io::Error> {
-        let dkim_service_url = env::var("DKIM_SERVICE_URL").map_err(|_| {
-            std::io::Error::new(std::io::ErrorKind::NotFound, "DKIM_SERVICE_URL not set")
-        })?;
-        let client = reqwest::Client::new();
-
-        let response = client
-            .post(&dkim_service_url)
-            .json(&serde_json::json!({
-                "from": email.from,
-                "to": email.to,
-                "subject": email.subject,
-                "text": email.body,
-                "html": email.body
-            }))
-            .send()
-            .await
-            .map_err(|e| std::io::Error::new(std::io::ErrorKind::Other, e))?;
-
-        let status = response.status();
-        let body = response
-            .text()
-            .await
-            .map_err(|e| std::io::Error::new(std::io::ErrorKind::Other, e))?;
-
-        if status.is_success() {
-            serde_json::from_str::<serde_json::Value>(&body)
-                .map_err(|e| std::io::Error::new(std::io::ErrorKind::Other, e))
-        } else {
-            let snippet = if body.len() > 1200 {
-                &body[..1200]
-            } else {
-                &body
-            };
-            Err(std::io::Error::new(
-                std::io::ErrorKind::Other,
-                format!("DKIM service HTTP {}: {}", status.as_u16(), snippet),
-            ))
-        }
-    }
-}
+// impl DkimService → dkim_service.rs
